refactor: Moved code block querying out
