+++ src/Visualisations/CBJuggl.ts
@@ -1,0 +1,5 @@
+import type {ParsedCodeblock} from "../interfaces";
+
+export function createdJugglCB(target: HTMLElement, args: ParsedCodeblock) {
+
+}
